Resolves gripper out of sync with arm motion
Trac ticket #11622

A customer reported that gripper playback was
not synchronized with the arm motion when playing
back a trajectory using the joint_trajectory_action_server
and joint_trajectory_file_playback (client).

This fix waits until feedback is recieved from the
joint trajectory action server before allowing the
playback of the grippers to begin.