Re #42 Added initial offset to playback wait
The in the joint_trajectory_file_playback example,
When a trajectory is recorded with the endpoint far
from the starting point, the trajectory did not fully
complete in playback. It appeared that playback was
cut short by approximately the amount of time it took
to reach the starting point when playback was initiated.

This change adds the initial time require to get to the
starting position into the overall trajectory's wait
calculation.

Resolves Trac #11978